+++ src/robotlegs/bender/framework/impl/ClassMatcher.ts
@@ -1,0 +1,24 @@
+// ------------------------------------------------------------------------------
+//  Copyright (c) 2017 RobotlegsJS. All Rights Reserved.
+//
+//  NOTICE: You are permitted to use, modify, and distribute this file
+//  in accordance with the terms of the license agreement accompanying it.
+// ------------------------------------------------------------------------------
+
+import { IMatcher } from "../api/IMatcher";
+
+/**
+ * @private
+ */
+export class ClassMatcher implements IMatcher {
+    /*============================================================================*/
+    /* Public Functions                                                           */
+    /*============================================================================*/
+
+    /**
+     * @private
+     */
+    public matches(item: any): boolean {
+        return typeof item === "function";
+    }
+}