+++ src/robotlegs/bender/framework/impl/ObjectProcessor.ts
@@ -7,6 +7,8 @@
 
 import { IMatcher } from "../api/IMatcher";
 
+import { ObjectHandler } from "./ObjectHandler";
+
 /**
  * Robotlegs object processor
  *
@@ -17,7 +19,7 @@ export class ObjectProcessor {
     /* Private Properties                                                         */
     /*============================================================================*/
 
-    private _handlers: any[] = [];
+    private _handlers: ObjectHandler[] = [];
 
     /*============================================================================*/
     /* Public Functions                                                           */
@@ -37,51 +39,15 @@ export class ObjectProcessor {
      * @param object The object instance to process.
      */
     public processObject(object: any): void {
-        for (let i: number = 0; i < this._handlers.length; i++) {
-            let handler: ObjectHandler = this._handlers[i];
+        this._handlers.forEach((handler: ObjectHandler) => {
             handler.handle(object);
-        }
+        });
     }
 
     /**
      * Removes all handlers
      */
     public removeAllHandlers(): void {
-        this._handlers.length = 0;
-    }
-}
-
-class ObjectHandler {
-    /*============================================================================*/
-    /* Private Properties                                                         */
-    /*============================================================================*/
-
-    private _matcher: IMatcher;
-
-    private _handler: Function;
-
-    /*============================================================================*/
-    /* Constructor                                                                */
-    /*============================================================================*/
-
-    /**
-     * @private
-     */
-    constructor(matcher: IMatcher, handler: Function) {
-        this._matcher = matcher;
-        this._handler = handler;
-    }
-
-    /*============================================================================*/
-    /* Public Functions                                                           */
-    /*============================================================================*/
-
-    /**
-     * @private
-     */
-    public handle(object: any): void {
-        if (this._matcher.matches(object)) {
-            this._handler(object);
-        }
+        this._handlers = [];
     }
 }