+++ test/entry.ts
@@ -1,6 +1,6 @@
 /// <reference types="mocha" />
 
-import "jsdom-global/register";
-import "bluebird/js/browser/bluebird";
-import "es6-symbol/implement";
-import "es6-map/implement";
+import "jsdom-global";
+import "bluebird";
+import "es6-symbol";
+import "es6-map";