+++ CHANGELOG.md
@@ -22,13 +22,81 @@
 
 - [ ] Migrate [original documentation](https://github.com/robotlegs/robotlegs-framework/blob/master/src/readme.md) and adapt it to TypeScript.
 
-## Robotlegs-Core 0.1.0
+## Robotlegs-Core 0.2.0
 
-### v0.1.4
+### v0.2.0
+
+Major Breaking Changes:
+---
+
+- **IEvent** interface changed to remove usage of **IEventInit** interface (see #57).
+
+    - Interface **IEvent** was:
+    ```typescript
+    export interface IEvent {
+        type: string;
+        defaultPrevented?: boolean;
+        bubbles?: boolean;
+        target?: any;
+        currentTarget?: any;
+        detail?: any;
+    }
+    ```
+
+    - Interface **IEvent** is now:
+    ```typescript
+    export interface IEvent {
+        type: string;
+        bubbles?: boolean;
+        cancelable?: boolean;
+        isDefaultPrevented?: boolean;
+        isPropagationStopped?: boolean;
+        isPropagationImmediateStopped?: boolean;
+        currentTarget?: any;
+        target?: any;
+        data?: any;
+        preventDefault(): void;
+        stopPropagation(): void;
+        stopImmediatePropagation(): void;
+    }
+    ```
+
+- Constructor of **Event** class changed to remove usage of **IEventInit** interface (see #57).
+
+    - Constructor of **Event** class was:
+    ```typescript
+    export class Event implements IEvent {
+        constructor(type: string, eventInit: IEventInit = { bubbles: false }) {
+            this.type = type;
+            this.defaultPrevented = false;
+            this.bubbles = eventInit.bubbles;
+            this.detail = eventInit.detail;
+        }
+    }
+    ```
+
+    - Constructor of **Event** is now:
+    ```typescript
+    export class Event implements IEvent {
+        constructor(type: string, bubbles?: boolean, cancelable?: boolean, data?: any) {
+            this._type = type;
+            this._bubbles = !!bubbles;
+            this._cancelable = !!cancelable;
+            this._data = data;
+        }
+    }
+    ```
+
+Features Or Improvements:
+---
+
+- Enforce TSLint rules (see #57).
 
 - Update TypeScript Compiler Options (see #56).
 
 - Update dev dependencies to latest version.
+
+## Robotlegs-Core 0.1.0
 
 ### [v0.1.3](https://github.com/RobotlegsJS/RobotlegsJS/releases/tag/0.1.3) - 2018-03-04